Refactor : Cleanup and delete of unused DTO's
