--- src/Application/Workflows/Dto/ArtifactStorageDto.cs
@@ -1,8 +1,0 @@
-﻿namespace Application.Workflows.Dto;
-
-public class ArtifactStorageDto(string key, string content)
-{
-    public string Key { get; } = key;
-
-    public string Content { get; } = content;
-}
