--- src/Application/Workflows/Dto/CreatePlanRequest.cs
@@ -1,8 +1,0 @@
-﻿using Application.Models;
-
-namespace Application.Workflows.Dto;
-
-public class CreatePlanRequestDto(TravelPlan travelPlan)
-{
-    public TravelPlan TravelPlan => travelPlan;
-}
