--- src/Application/Workflows/Dto/FlightDto.cs
@@ -1,29 +1,0 @@
-﻿namespace Application.Workflows.Dto;
-
-public class FlightSearchResultDto
-{
-    public string ArtifactKey { get; set; }
-    public List<FlightOptionDto> Results { get; set; }
-}
-
-public class FlightOptionDto
-{
-    public string Airline { get; set; }
-    public string FlightNumber { get; set; }
-    public FlightEndpointDto Departure { get; set; }
-    public FlightEndpointDto Arrival { get; set; }
-    public string Duration { get; set; }
-    public PriceDto Price { get; set; }
-}
-
-public class FlightEndpointDto
-{
-    public string Airport { get; set; }
-    public DateTime Datetime { get; set; }
-}
-
-public class PriceDto
-{
-    public decimal Amount { get; set; }
-    public string Currency { get; set; }
-}
