--- src/Application/Workflows/Dto/HotelDto.cs
@@ -1,24 +1,0 @@
-namespace Application.Workflows.Dto;
-
-public record HotelSearchResultDto
-{
-    public string ArtifactKey { get; init; }
-    public List<HotelOptionDto> Results { get; init; }
-}
-
-public record HotelOptionDto
-{
-    public string HotelName { get; init; }
-    public string Address { get; init; }
-    public DateTime CheckIn { get; init; }
-    public DateTime CheckOut { get; init; }
-    public decimal Rating { get; init; }
-    public HotelPriceDto PricePerNight { get; init; }
-    public HotelPriceDto TotalPrice { get; init; }
-}
-
-public class HotelPriceDto
-{
-    public decimal Amount { get; set; }
-    public string Currency { get; set; }
-}
