--- src/Application/Workflows/Dto/WorkflowResponse.cs
@@ -1,7 +1,0 @@
-﻿namespace Application.Workflows.Dto;
-
-public class WorkflowResponse(WorkflowState state, string message)
-{
-    public WorkflowState State { get; } = state;
-    public string Message { get; } = message;
-}
