--- src/Application/Workflows/Dto/WorkflowStateDto.cs
@@ -1,10 +1,0 @@
-﻿using Microsoft.Agents.AI.Workflows;
-
-namespace Application.Workflows.Dto;
-
-public class WorkflowStateDto(WorkflowState state, CheckpointInfo? checkpointInfo)
-{
-    public WorkflowState State { get; } = state;
-
-    public CheckpointInfo? CheckpointInfo { get; } = checkpointInfo;
-}
